fix problem with training data
