+++ .gitignore
@@ -317,3 +317,22 @@ input/EUROMODoutput/sc expanded/uk_2024_std.txt
 input/EUROMODoutput/sc expanded/uk_2025_std.txt
 input/EUROMODoutput/sc expanded/uk_2026_std.txt
 input/EUROMODoutput/sc expanded/uk_2027_std.txt
+input/EUROMODoutput/current/DatabaseCountryYear.xlsx
+input/EUROMODoutput/current/EUROMODpolicySchedule.xlsx
+input/EUROMODoutput/current/uk_2011_std.txt
+input/EUROMODoutput/current/uk_2012_std.txt
+input/EUROMODoutput/current/uk_2013_std.txt
+input/EUROMODoutput/current/uk_2014_std.txt
+input/EUROMODoutput/current/uk_2015_std.txt
+input/EUROMODoutput/current/uk_2016_std.txt
+input/EUROMODoutput/current/uk_2017_std.txt
+input/EUROMODoutput/current/uk_2018_std.txt
+input/EUROMODoutput/current/uk_2019_std.txt
+input/EUROMODoutput/current/uk_2020_std.txt
+input/EUROMODoutput/current/uk_2021_std.txt
+input/EUROMODoutput/current/uk_2022_std.txt
+input/EUROMODoutput/current/uk_2023_std.txt
+input/EUROMODoutput/current/uk_2024_std.txt
+input/EUROMODoutput/current/uk_2025_std.txt
+input/EUROMODoutput/current/uk_2026_std.txt
+input/EUROMODoutput/current/uk_2027_std.txt
